backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27.2
feedparser==6.0.11
langdetect==1.0.9
jieba==0.42.1
pgvector==0.3.6
apscheduler==3.10.4
crawl4ai==0.4.248
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0
